// src/features/pengaturan-tegangan/types.ts
export interface PengaturanTeganganList {
  gardu_induk: {
    id: string;
    nama: string;
  };
  id: string;
  jenis: string;
  jurusan: string;
  jurusan_id: string;
  keterangan: string;
  mvar: number;
  open_close: string;
  sebelum: number;
  sesudah: number;
  tanggal: string;
  waktu: string;
}